RAD-224 Extract RadiologyStudyService
* move Study to own package
* extract RadiologyStudyService into it
* move hibernate files to study package
* make HibernateStudyDAO StudyDAO package private
* create own RadiologyStudyServiceComponentTestDataset.xml
