+++ api/src/main/java/org/openmrs/module/radiology/study/Study.java
@@ -7,9 +7,15 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology;
+package org.openmrs.module.radiology.study;
 
 import java.lang.reflect.Field;
+
+import org.openmrs.module.radiology.Modality;
+import org.openmrs.module.radiology.MwlStatus;
+import org.openmrs.module.radiology.PerformedProcedureStepStatus;
+import org.openmrs.module.radiology.RadiologyOrder;
+import org.openmrs.module.radiology.ScheduledProcedureStepStatus;
 
 /**
  * A class that supports on openmrs's orders to make the module DICOM compatible, corresponds to the
@@ -99,7 +105,7 @@ public class Study {
 		this.modality = modality;
 	}
 	
-	void setRadiologyOrder(RadiologyOrder radiologyOrder) {
+	public void setRadiologyOrder(RadiologyOrder radiologyOrder) {
 		this.radiologyOrder = radiologyOrder;
 	}
