+++ api/src/main/resources/RadiologyStudy.hbm.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN" "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
 
 <hibernate-mapping package="org.openmrs.module.radiology">
-	<class name="Study" table="radiology_study">
+	<class name="org.openmrs.module.radiology.study.Study" table="radiology_study">
 		<id name="studyId" type="int" column="study_id" unsaved-value="0">
 			<generator class="native">
 				<param name="sequence">radiology_study_study_id_seq</param>
