+++ api/src/main/resources/moduleApplicationContext.xml
@@ -41,6 +41,33 @@
 		</property>
 	</bean>
 
+	<bean id="radiologyStudyService"
+		class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+		<property name="transactionManager">
+			<ref bean="transactionManager" />
+		</property>
+		<property name="target">
+			<bean
+				class="org.openmrs.module.radiology.study.RadiologyStudyServiceImpl">
+			</bean>
+		</property>
+		<property name="preInterceptors">
+			<ref bean="serviceInterceptors" />
+		</property>
+		<property name="transactionAttributeSource">
+			<ref bean="transactionAttributeSource" />
+		</property>
+	</bean>
+
+	<bean parent="serviceContext">
+		<property name="moduleService">
+			<list merge="true">
+				<value>org.openmrs.module.radiology.study.RadiologyStudyService</value>
+				<ref local="radiologyStudyService" />
+			</list>
+		</property>
+	</bean>
+
 	<bean id="radiologyReportService"
 		class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
 		<property name="transactionManager">
