+++ api/src/test/java/org/openmrs/module/radiology/RadiologyServiceComponentTest.java
@@ -23,7 +23,6 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -46,6 +45,7 @@ import org.openmrs.api.OrderService;
 import org.openmrs.api.PatientService;
 import org.openmrs.api.ProviderService;
 import org.openmrs.api.VisitService;
+import org.openmrs.module.radiology.study.Study;
 import org.openmrs.test.BaseModuleContextSensitiveTest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -715,211 +715,5 @@ public class RadiologyServiceComponentTest extends BaseModuleContextSensitiveTes
 		expectedException.expect(IllegalArgumentException.class);
 		expectedException.expectMessage("radiologyOrder is not persisted");
 		radiologyService.discontinueRadiologyOrderInPacs(new RadiologyOrder());
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByStudyId(Integer)
-	 * @verifies should return study for given study id
-	 */
-	@Test
-	public void getStudyByStudyId_shouldReturnStudyForGivenStudyId() throws Exception {
-		
-		Study study = radiologyService.getStudyByStudyId(EXISTING_STUDY_ID);
-		
-		assertNotNull(study);
-		assertThat(study.getRadiologyOrder()
-				.getOrderId(), is(EXISTING_RADIOLOGY_ORDER_ID));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByStudyId(Integer)
-	 * @verifies should return null if no match was found
-	 */
-	@Test
-	public void getStudyByStudyId_shouldReturnNullIfNoMatchIsFound() throws Exception {
-		
-		Study study = radiologyService.getStudyByStudyId(NON_EXISTING_STUDY_ID);
-		
-		assertNull(study);
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByOrderId(Integer)
-	 * @verifies should return study associated with radiology order for which order id is given
-	 */
-	@Test
-	public void getStudyByOrderId_shouldReturnStudyMatching() throws Exception {
-		
-		Study study = radiologyService.getStudyByOrderId(EXISTING_RADIOLOGY_ORDER_ID);
-		
-		assertNotNull(study);
-		assertThat(study.getRadiologyOrder()
-				.getOrderId(), is(EXISTING_RADIOLOGY_ORDER_ID));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByOrderId(Integer)
-	 * @verifies should return null if no match was found
-	 */
-	@Test
-	public void getStudyByOrderId_shouldReturnNullIfNoMatchIsFound() {
-		
-		Study study = radiologyService.getStudyByOrderId(NON_EXISTING_RADIOLOGY_ORDER_ID);
-		
-		assertNull(study);
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByOrderId(Integer)
-	 * @verifies should throw illegal argument exception given null
-	 */
-	@Test
-	public void getStudyByOrderId_shouldThrowIllegalArgumentExceptionGivenNull() {
-		
-		expectedException.expect(IllegalArgumentException.class);
-		expectedException.expectMessage("orderId is required");
-		radiologyService.getStudyByOrderId(null);
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByStudyInstanceUid(String)
-	 * @verifies should return study matching study instance uid
-	 */
-	@Test
-	public void getStudyByStudyInstanceUid_shouldReturnStudyMatchingUid() throws Exception {
-		Study study = radiologyService.getStudyByStudyInstanceUid(EXISTING_STUDY_INSTANCE_UID);
-		
-		assertNotNull(study);
-		assertThat(study.getStudyInstanceUid(), is(EXISTING_STUDY_INSTANCE_UID));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByStudyInstanceUid(String)
-	 * @verifies should return null if no match was found
-	 */
-	@Test
-	public void getStudyByStudyInstanceUid_shouldReturnNullIfNoMatchIsFound() throws Exception {
-		Study study = radiologyService.getStudyByStudyInstanceUid(NON_EXISTING_STUDY_INSTANCE_UID);
-		
-		assertNull(study);
-	}
-	
-	/**
-	 * @see RadiologyService#getStudyByStudyInstanceUid(String)
-	 * @verifies should throw IllegalArgumentException if study instance uid is null
-	 */
-	@Test
-	public void getStudyByStudyInstanceUid_shouldThrowIllegalArgumentExceptionIfUidIsNull() throws Exception {
-		expectedException.expect(IllegalArgumentException.class);
-		expectedException.expectMessage("studyInstanceUid is required");
-		radiologyService.getStudyByStudyInstanceUid(null);
-	}
-	
-	/**
-	 * @see RadiologyService#getStudiesByRadiologyOrders(List<RadiologyOrder>)
-	 * @verifies should fetch all studies for given radiology orders
-	 */
-	@Test
-	public void getStudiesByRadiologyOrders_shouldFetchAllStudiesForGivenRadiologyOrders() throws Exception {
-		
-		Patient patient = patientService.getPatient(PATIENT_ID_WITH_TWO_STUDIES_AND_NO_NON_RADIOLOGY_ORDER);
-		List<RadiologyOrder> radiologyOrders = radiologyService.getRadiologyOrdersByPatient(patient);
-		
-		List<Study> studies = radiologyService.getStudiesByRadiologyOrders(radiologyOrders);
-		
-		assertThat(studies.size(), is(radiologyOrders.size()));
-		assertThat(studies.get(0)
-				.getRadiologyOrder(), is(radiologyOrders.get(0)));
-		assertThat(studies.get(1)
-				.getRadiologyOrder(), is(radiologyOrders.get(1)));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudiesByRadiologyOrders(List<RadiologyOrder>)
-	 * @verifies should return empty list given radiology orders without associated studies
-	 */
-	@Test
-	public void getStudiesByRadiologyOrders_shouldReturnEmptyListGivenRadiologyOrdersWithoutAssociatedStudies()
-			throws Exception {
-		
-		RadiologyOrder radiologyOrderWithoutStudy = radiologyService.getRadiologyOrderByOrderId(RADIOLOGY_ORDER_ID_WITHOUT_STUDY);
-		List<RadiologyOrder> radiologyOrders = Arrays.asList(radiologyOrderWithoutStudy);
-		
-		List<Study> studies = radiologyService.getStudiesByRadiologyOrders(radiologyOrders);
-		
-		assertThat(radiologyOrders.size(), is(1));
-		assertThat(studies.size(), is(0));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudiesByRadiologyOrders(List<RadiologyOrder>)
-	 * @verifies should return empty list given empty radiology order list
-	 */
-	@Test
-	public void getStudiesByRadiologyOrders_shouldReturnEmptyListGivenEmptyRadiologyOrderList() throws Exception {
-		
-		List<RadiologyOrder> orders = new ArrayList<RadiologyOrder>();
-		
-		List<Study> studies = radiologyService.getStudiesByRadiologyOrders(orders);
-		
-		assertThat(orders.size(), is(0));
-		assertThat(studies.size(), is(0));
-	}
-	
-	/**
-	 * @see RadiologyService#getStudiesByRadiologyOrders(List<RadiologyOrder>)
-	 * @verifies should throw IllegalArgumentException given null
-	 */
-	@Test
-	public void getStudiesByRadiologyOrders_shouldThrowIllegalArgumentExceptionGivenNull() throws Exception {
-		
-		expectedException.expect(IllegalArgumentException.class);
-		expectedException.expectMessage("radiologyOrders are required");
-		radiologyService.getStudiesByRadiologyOrders(null);
-	}
-	
-	/**
-	 * @see RadiologyService#updateStudyPerformedStatus(String,PerformedProcedureStepStatus)
-	 * @verifies update performed status of study associated with given study instance uid
-	 */
-	@Test
-	public void updateStudyPerformedStatus_shouldUpdatePerformedStatusOfStudyAssociatedWithGivenStudyInstanceUid()
-			throws Exception {
-		
-		Study existingStudy = radiologyService.getStudyByStudyId(EXISTING_STUDY_ID);
-		PerformedProcedureStepStatus performedStatusPreUpdate = existingStudy.getPerformedStatus();
-		PerformedProcedureStepStatus performedStatusPostUpdate = PerformedProcedureStepStatus.COMPLETED;
-		
-		Study updatedStudy = radiologyService.updateStudyPerformedStatus(existingStudy.getStudyInstanceUid(),
-			performedStatusPostUpdate);
-		
-		assertNotNull(updatedStudy);
-		assertThat(updatedStudy, is(existingStudy));
-		assertThat(performedStatusPreUpdate, is(not(performedStatusPostUpdate)));
-		assertThat(updatedStudy.getPerformedStatus(), is(performedStatusPostUpdate));
-	}
-	
-	/**
-	 * @see RadiologyService#updateStudyPerformedStatus(String,PerformedProcedureStepStatus)
-	 * @verifies throw illegal argument exception if study instance uid is null
-	 */
-	@Test
-	public void updateStudyPerformedStatus_shouldThrowIllegalArgumentExceptionIfStudyInstanceUidIsNull() throws Exception {
-		
-		expectedException.expect(IllegalArgumentException.class);
-		expectedException.expectMessage("studyInstanceUid is required");
-		radiologyService.updateStudyPerformedStatus(null, PerformedProcedureStepStatus.COMPLETED);
-	}
-	
-	/**
-	 * @see RadiologyService#updateStudyPerformedStatus(String,PerformedProcedureStepStatus)
-	 * @verifies throw illegal argument exception if performed status is null
-	 */
-	@Test
-	public void updateStudyPerformedStatus_shouldThrowIllegalArgumentExceptionIfPerformedStatusIsNull() throws Exception {
-		
-		expectedException.expect(IllegalArgumentException.class);
-		expectedException.expectMessage("performedStatus is required");
-		radiologyService.updateStudyPerformedStatus(EXISTING_STUDY_INSTANCE_UID, null);
 	}
 }
