+++ api/src/test/java/org/openmrs/module/radiology/study/StudyTest.java
@@ -1,10 +1,14 @@
-package org.openmrs.module.radiology;
+package org.openmrs.module.radiology.study;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+import org.openmrs.module.radiology.PerformedProcedureStepStatus;
 
+/**
+ * Tests {@link Study}.
+ */
 public class StudyTest {
 	
 	/**
