+++ omod/src/test/java/org/openmrs/module/radiology/test/RadiologyTestData.java
@@ -36,8 +36,8 @@ import org.openmrs.User;
 import org.openmrs.module.radiology.Modality;
 import org.openmrs.module.radiology.PerformedProcedureStepStatus;
 import org.openmrs.module.radiology.RadiologyOrder;
-import org.openmrs.module.radiology.Study;
 import org.openmrs.module.radiology.report.RadiologyReport;
+import org.openmrs.module.radiology.study.Study;
 import org.openmrs.util.RoleConstants;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.mock.web.MockMultipartHttpServletRequest;
